requirement.txt
absl-py==2.1.0
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
diffusers==0.32.0
dill==0.3.8
einops==0.8.0
etaprogress==1.1.1
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
fvcore==0.1.5.post20221221
grpcio==1.68.0
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
iopath==0.1.10
Jinja2==3.1.4
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.3.1
multiprocess==0.70.16
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==3.0.0
propcache==0.3.1
protobuf==5.28.3
psutil==7.0.0
ptflops==0.7.4
pthflops==0.4.2
pyarrow==19.0.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
thop==0.1.1.post2209072238
tifffile==2024.9.20
torch==2.5.1
torch-tb-profiler==0.4.3
torchaudio==2.5.1
torchinfo==1.8.0
torchsummary==1.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0
